Update rawchannel.cc
Before switch async_receiver_ready() only with the canFD_struct. 
Test will be performed with a mix of device in same time. MCP2515 for CAN_HS and PEAK CAN_FD together.
Goal will be to check how react the driver when a device is not compliant with CAN_FD. The canFD_struct should be compatible with the 2 drivers (cf can-utils/candump.c) but some test must be done before.
I will update the code only after this last integration test.